+++ changelog.md
@@ -2,13 +2,19 @@
 
 > All Time Stamps: GMT+8
 
-## 0.3.0.2 _2020-05-21 17:00_
+## 0.3.1.0 _2020-05-27 22:30_
+
+* 加入材料 added materials
+  * 激流水晶 torrential crystal
+* 加入特性 added traits
+  * 扰动(护甲) turbulent(armor)
+
+## 0.3.0.2 _2020-05-26 17:00_
 
 * 加入方块和物品 added blocks and items
   * 硬实泥土矿 solid dirt ore block
   * 硬实沙矿 solid sand ore block
   * 硬实石头矿 solid stone ore block
-  * 叶绿粉 chloroplast dust item
   * 彩色玻璃块 colorful glass block
   * 激流推进器 torrential thruster
   * 导管 tube item